docker/flask/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.6
gunicorn==23.0.0
werkzeug>=2.3.0
